Create Day 4.3
Chef buys a lottery ticket that has a 3 digit code associated with it. He thinks that digit 7 is his lucky digit and brings him good luck. Chef will win some amount in the lottery if at least one of the digits of the lottery ticket is 7.

Given three digits A, B, and C of the lottery ticket, tell whether Chef wins something or not?

Input Format
First line will contain T, the number of test cases. Then the test cases follow.
Each test case contains a single line of input, three space separated integers A,B,C.
Output Format
For each testcase, output in a single line answer "YES" if Chef wins a positive amount with the lottery and "NO" if not.